Update: timeout mechanism
The repository data scraper module has been updated with the 'BadObject' import, which was previously missing. Additionally, a timeout feature has been implemented in the process of mining additional cherry-pick scenarios from repository data. This feature ensures that no more than a specified amount of time (in this case, 3 minutes) is spent on retrieving excessive amounts of scenarios from a single repository.