Add 'repository_name' field to RepositoryDataScraper
A new field 'repository_name' is introduced in the initializer of class RepositoryDataScraper and used in the progress report of the scraping process. This helps provide more identifiable information in the progress output, in place of using the repository's full origin URL, which is unavailable in the local demo repositories also available in the project repo.